SSM-4927 Implemented Test button to validate query.
git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@41808 219f1c6e-4a0e-43c8-96f2-6e749be13fd4

+++ modules/gateway/common/src/main/java/com/l7tech/gateway/common/cassandra/CassandraConnectionManagerAdmin.java
@@ -88,5 +88,5 @@ public interface CassandraConnectionManagerAdmin extends AsyncAdminMethods {
      */
     @Transactional(readOnly=true)
     @Secured(types = EntityType.CASSANDRA_CONFIGURATION, stereotype = MethodStereotype.TEST_CONFIGURATION)
-    AsyncAdminMethods.JobId<String> testCassandraQuery(String connectionName, String query, int queryTimeout);
+    AsyncAdminMethods.JobId<String> testCassandraQuery(String connectionName, String query, long queryTimeout);
 }
